+++ src/hexdoc/jinja/__init__.py
@@ -2,6 +2,7 @@ __all__ = [
     "IncludeRawExtension",
     "hexdoc_item_image",
     "hexdoc_localize",
+    "hexdoc_smart_var",
     "hexdoc_texture_image",
     "hexdoc_wrap",
 ]
@@ -10,6 +11,7 @@ from .extensions import IncludeRawExtension
 from .filters import (
     hexdoc_item_image,
     hexdoc_localize,
+    hexdoc_smart_var,
     hexdoc_texture_image,
     hexdoc_wrap,
 )