+++ src/hexdoc/jinja/render.py
@@ -28,6 +28,7 @@ from .extensions import DefaultMacroExtension, IncludeRawExtension
 from .filters import (
     hexdoc_item_image,
     hexdoc_localize,
+    hexdoc_smart_var,
     hexdoc_texture_image,
     hexdoc_wrap,
 )
@@ -107,6 +108,7 @@ def create_jinja_env_with_loader(loader: BaseLoader):
         "hexdoc_localize": hexdoc_localize,
         "hexdoc_texture_image": hexdoc_texture_image,
         "hexdoc_item_image": hexdoc_item_image,
+        "hexdoc_smart_var": hexdoc_smart_var,
     }
 
     return env
@@ -193,6 +195,7 @@ def render_book(
         "site_url": str(props.env.github_pages_url),
         "relative_site_url": relative_site_url,
         "page_url": str(page_url),
+        "source_url": str(props.env.source_url),
         "version": version,
         "lang": lang,
         "lang_name": lang_name,
@@ -204,6 +207,11 @@ def render_book(
         "safari_pinned_tab_color": "#332233",
         "minecraft_version": minecraft_version or "???",
         "full_version": plugin.full_version,
+        "navbar": {  # default navbar links (ignored if set in props)
+            "center": [
+                {"text": "GitHub", "href": {"variable": "source_url"}},
+            ],
+        },
         "_": lambda key: hexdoc_localize(  # i18n helper
             key,
             do_format=False,