+++ test/patchouli/test_text.py
@@ -281,7 +281,7 @@ def test_case_insensitive_entry_id():
         link_overrides={},
     )
 
-    href = style.href({"book_links": {"items/staff": URL("link")}})
+    href = style.href({"book_links": {"namespace:items/staff": URL("link")}})
 
     assert href == "link"
 
@@ -293,7 +293,7 @@ def test_case_sensitive_page_anchor_match():
         link_overrides={},
     )
 
-    href = style.href({"book_links": {"items/staff#anchor": URL("link")}})
+    href = style.href({"book_links": {"namespace:items/staff#anchor": URL("link")}})
 
     assert href == "link"
 
@@ -306,4 +306,4 @@ def test_case_sensitive_page_anchor_broken():
     )
 
     with pytest.raises(ValueError):
-        style.href({"book_links": {"items/staff#anchor": URL("link")}})
+        style.href({"book_links": {"namespace:items/staff#anchor": URL("link")}})